requirements.txt
flask==3.1.3
flask-cors==6.0.2
opencv-python==4.13.0.92
mediapipe==0.10.14
numpy==2.4.2
Pillow==12.1.1
werkzeug==3.1.6
